+++ src/crypto/mode/cfb.js
@@ -18,15 +18,15 @@
  */
 
 /**
- * @module crypto/cfb
+ * @module crypto/mode/cfb
  * @private
  */
 
 import { AES_CFB } from '@openpgp/asmcrypto.js/dist_es8/aes/cfb';
 
 import stream from '@openpgp/web-stream-tools';
-import * as cipher from './cipher';
-import util from '../util';
+import * as cipher from '../cipher';
+import util from '../../util';
 
 const webCrypto = util.getWebCrypto();
 const nodeCrypto = util.getNodeCrypto();
